localisation_hardware/requirements.txt
absl-py==0.8.0
astor==0.8.0
audioread==2.1.8
cffi==1.12.3
decorator==4.4.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
joblib==0.13.2
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
numba==0.45.1
numpy==1.17.2
pandas==0.25.1
protobuf==3.9.1
PyAudio==0.2.11
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
resampy==0.2.2
RPi.GPIO==0.7.0
rpimotorlib==2.7
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
SoundFile==0.10.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.36.1
Werkzeug==0.16.0
wrapt==1.11.2
